Use unique mock urls for each test. Also fix a bug where a PaymentOrderUrls initializer ignored the identifier argument.
URLProtocol is used to mock server responses. It was possible that requests to mock urls of a previous test would be answered by stubs of a subsequent test. Rather than try to painstakingly synchronize the requests between tests, we can simply use unique mock urls per test to get rid of this problem.